Extract the function `openPickedProject` and reuse in Side Bar
